+++ R/null.test_fun.R
@@ -22,7 +22,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
                                     arguments = null.args,
                                     cor.matrix = null.cor,
                                     scree = null.scree),
-                            metric = metric, dimensions = length(data$call$dimensions)),
+                            metric = metric, dimensions = data$call$dimensions),
                     cent.tend = mean, quantiles = 1)$obs)
         } else {
                 null_models_result <- replicate(replicates, summary(dispRity(
@@ -32,7 +32,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
                                     arguments = null.args,
                                     cor.matrix = null.cor,
                                     scree = null.scree),
-                            metric = metric, dimensions = length(data$call$dimensions), args),
+                            metric = metric, dimensions = data$call$dimensions, args),
                     cent.tend = mean, quantiles = 1)$obs)        }
     } else {
         if(is.null(args)) {
@@ -45,7 +45,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
                                     cor.matrix = null.cor,
                                     scree = null.scree)
                         ),
-                    metric = metric, dimensions = length(data$call$dimensions)),
+                    metric = metric, dimensions = data$call$dimensions),
             cent.tend = mean, quantiles = 1)$obs)
         } else {
             null_models_result <- replicate(replicates, summary(dispRity(
@@ -57,10 +57,22 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
                                     cor.matrix = null.cor,
                                     scree = null.scree)
                         ),
-                    metric = metric, dimensions = length(data$call$dimensions), args),
+                    metric = metric, dimensions = data$call$dimensions, args),
             cent.tend = mean, quantiles = 1)$obs)
         }
     }
     options(warn = 0)
     return(null_models_result)
 }
+
+
+# unlist(extract.dispRity(dispRity(
+#                     scale(
+#                         space.maker(nrow(data$matrix[[1]]),
+#                                     dimensions = length(data$call$dimensions),
+#                                     distribution = null.distrib,
+#                                     arguments = null.args,
+#                                     cor.matrix = null.cor,
+#                                     scree = null.scree)
+#                         ),
+#                     metric = metric, dimensions = data$call$dimensions))))