+++ tests/testthat/test-null.test.R
@@ -108,7 +108,9 @@ test_that("make.null.model works", {
         , test3)
 
     set.seed(1)
-    expect_warning(test5 <- make.null.model(single_disp, replicates = 5, null.distrib = my_distributions1, null.args = NULL, null.cor = NULL, scale = TRUE, null.scree = NULL, metric = centroids, args = 0))
+    warning <- capture_warnings(test5 <- make.null.model(single_disp, replicates = 5, null.distrib = my_distributions1, null.args = NULL, null.cor = NULL, scale = TRUE, null.scree = NULL, metric = centroids, args = 0))
+    expect_equal(warning[[1]], "Row names have been automatically added to data.")
+    # expect_equal(warning[[6]], "The metric applied on the null models is a dimensions level 2 metric. Results are averaged.")
     expect_is(test5, "numeric")
     expect_equal(length(test5), 5)
 })